Refactor IsAuthenticated logic to ensure valid JSON return
We had some problems when error messages had double quotes in them, as
it would break the JSON unmarshaling. To avoid this kind of errors
again, we now use separate types for the returned data and marshal them
through json.Marshal, which prevents mistakes that could happen when
"manually" generating the message.