No longer run daemon tests in parallel
After the changes introduced in #52, the broker bus is always exported
with the same name. This means that we can't run multiple instances of
the daemon as they would fail when competing for the same bus name.
Since we can't run multiple daemons in production as well (for the same
reason), it's fine to drop the parallel testing and run them
sequentially instead.