Allow to start broker and sessions in offline mode
Even though we "allowed" offline authentication, the broker would not
start without a network connection. Now, the broker does not rely on a
connection to run and now sessions have control over the auth config
needed to authenticate the user online. This means we can start
sessions in offline mode when failing to connect to the provider.