Use more descriptive error messages
We used "could not get user info" in too many places, making it hard to
tell where this error message originated from.

Also, add a brief description to errors which originate in third-party
packages before passing them up the call stack, to make it easier to
tell where in our code we first encountered the error.