Better errors when failing to create broker object
We used to have a quite big prefix for the error and use the same
message for when either the client ID or the issuer URL were not
configured. Now, we can show a more targeted error that also won't
require the user to scroll right to fully see the message.