Revert "Update user info even on local authentication"
Commit efd871936e6e6be521f2337e7019aa0af216715e broke local
authentication when the access token is _not_ expired. With that commit,
we tried to get the groups with the locally stored token, which in case
of the MSEntraID provider, returns an error when the scopes field
doesn't contain the GroupMember.Read.All scope. The problem is that we
don't store the extra scopes field of the token locally, so it always
fails with:

    could not get user info: failed to cast token scopes to string: <nil>.

This reverts commit efd871936e6e6be521f2337e7019aa0af216715e.