Check that usernames comply with the Microsoft Entra username policy
We must avoid that multiple users of the provider map to the same local
user. We currently use `strings.EqualFold` to compare the usernames,
which considers some different Unicode characters the same (for example
the "KELVIN SIGN" (U+212A) and the ASCII "K"). To avoid that, we
considered using a custom `toLowercase` function which only converts
ASCII characters to lowercase. However, we can just check if the
username is valid under the Microsoft Entra username policy.

Microsoft Entra enforces the username policy for the User Principal Name
(UPN). We currently don't use the UPN as the username though, we use the
`preferred_username` field from the ID token. The value of that field is
equal to the UPN in some cases, in other cases it's equal to the user's
email address. It's not clear if the username policy is enforced for all
possible values of the `preferred_username` field.

So to ensure that we don't have to care about the Unicode case-folding
when comparing the usernames, we enforce the username policy ourselves.

UDENG-4517