Rework ID token generation for the tests
We used to ignore the JWT format in the tests, so we could manually
write the ID Token. Now that we decided to implement the JWT logic
also, the ID token structure becomes more complex:

{header}.{payload}.{signature_string}

To avoid generating all of this manually, it's better to switch to a
package that already handles this parsing with the right encryption and
encoding.