google specific provider
We need to override the generic provider (which should be renamed from
NoProvider to BaseProvider in the long term) to remove the
offline_access scope. For TV and limited input devices apps, Google APIs
will reject any request with this scope, while still allowing non
interactive user access token refresh.
All the rest of the functionalities is aligned with the NoProvider one.