ci: Use gotestfmt to format go test output
The output of go test is notoriously hard to read, especially when a lot
of tests are executed and a lot of output is produced, which is the case
in our CI.

This commit uses gotestfmt to format the go test output. It only prints
the output of failed tests to stdout, because that's what we're usually
interested in. In case we need the full test output, that's now uploaded
as an artifact.