fix: add device ownership migration logic
The device ownership config, added in
e2d7fe632517e6514bc91cdfb2ce126d0e04c37c, introduced a breaking change.
The default behavior before was that ALL users were allowed, but now only
the owner is allowed.

In order to migrate the config, we set a new snapctl config param. If
that is not present, that means that we need to migrate the config.