Hide --config flag
With 6b82eda02032232d71313092dc51eea59333ce5a, we automatically generate
config files in the `.d` directory of the config file. We don't want to
create those in `$PWD/mybroker.conf.d` when `--config mybroker.conf` is
used.

To avoid additional complexity, we decided to keep the current behavior
but hide the flag from the usage message, so that it's only used for
debugging and testing.