Fix get_version script
Building from main or any other branch that's not "msentraid" or
"google" failed because it used "notag" which is not a valid semantic
version.

It now first tries to use the highest version tag prefixed with the branch
name (e.g. "msentraid-0.2.0" when on the msentraid branch) and if there
is none, just uses the highest version tag on the branch which starts
with a number (to avoid an invalid SemVer when tag "msentraid-0.2.0" is
on a branch that's not "msentraid").