Fix snap releases still having version 0.2.0-pre1 instead of 0.2.0
`git tag --sort=-v:refname` sorts tagnames with the same base version
but different suffixes in lexicographical order, causing prerelease tags
to be listed as higher than the main release. This results in the snaps
still having `0.2.0-pre1+...` versions instead of `0.2.0+...`.

Git supports the `versionsort.suffix` option to avoid exactly that, so
let's use it.