refactor: Clean separation of auth modes supported by provider and UI
We used the auth modes supported by the UI to define the list of auth
modes supported by provider endpoints, which was confusing.

Also, print a log message when the UI doesn't support an authentication
mode which is supported by the provider, instead of returning an error,
because that's not an error if there are other modes available.