Add tags to ignore static check on helper files
We have a lot of helper files to help us test our project and those
are protected to not be built with "go build" to avoid having those in
the release binary.

Since they are not part of what we use in production, we don't want
them listed in the static analysis checks as it would give us wrong
metrics on the results.

To avoid having to update tics configuration every time we rename
those files or add new ones, let's tag them and filter those out in
the analysis.