pam/dbusmodule/transaction: Expose Context to monitor the DBus connection
When a dbus transaction is used we may want to be able to monitor if the
server has been disconnected and react accordingly.

We were mostly relying so far on the fact that calling a method would
have failed, but since we can easily monitor the connection state, let's
do that instead.