Improve error message shown when user is not allowed
Multiple users were confused by the default behavior that only the first
user is allowed to log in and all other users get a permission denied.

The documentation was improved in https://github.com/ubuntu/authd/pull/888.

This commit also improves the error message that's shown when the user
is not allowed, to explain better why the login failed.