Print a message when the refresh token is expired
I encountered the case that the refresh token in one of my VMs was
expired and was puzzled for a while why the PAM module auto-selects
device authentication after I entered my local password. Let's print a
message which makes it clear to the user why device authentication is
required.