Reduce risk of re-registering a device
If we register the device and then something goes wrong before we're
able to store the registration data on disk, the device will be
re-registered upon the next login, causing multiple registrations for
the same device in Entra ID.

We can't avoid that completely, but we can reduce the risk by storing
the registration data immediately after registering the device, so that
e.g. if the login times out during the "choose local password" prompt or
the user aborts the login, we already have the registration data stored
on disk.