e2e-tests: Make test cases not rely on other test cases
This makes `--rerunfailed` actually work. The issue was that
`--rerunfailed` only reruns *test cases* which failed. Since our test
suites contained multiple test cases which depended on each other, if
the first one passed but the second failed, `--rerunfailed` would only
start with the second one, which would then fail because the VM is not
in the expected state (because the first test case was skipped).

The quick fix is to only have one single test case per test suite. We
could still use multiple test cases per test suite if we ensure that
they don't depend on each other and the state of the VM is reset
before each test case (e.g. by restoring a snapshot in [Setup]).