e2e-tests/vm: Make noble bind sshd to VSOCK
With systemd v256, sshd automatically listens on VSOCK if there is a
VSOCK. Unfortunately, the systemd version on noble doesn't have that
feature yet, so we tried connecting to sshd via IP instead.

That is causing issues, for example sometimes `virsh domifaddr` and
`virsh net-dhcp-leases` don't show the domain's IP address, even though
the VM does have an IP address and internet connectivity.

This adds a simple systemd service to the noble VM which uses socat to
make sshd accessible to VSOCK.

Let's see if that makes things more robust.