e2e-tests/vm/provision.sh: Make rebooting more robust
I've seen the e2e tests fail because the system rebooted a few seconds
after restoring the ${broker}-edge-configured snapshot. One thing that
would explain that behavior is if the "Reboot VM and wait until it's
back" step succeeds after sending the VM the signal that it should
reboot but without waiting for it to actually reboot and then creating
the snapshot. Let's try to make that step more robust by ensuring that
the VM was actually shut down.