e2e-tests: Extract installation of deps into scripts
The test deps were installed in the provision-authd.sh script which
doesn't really make sense since they are not needed for provisioning.

They can now be executed separately from the provisioning scripts.

The testing.md now also refers to that script to avoid some duplication
of the list of dependencies (and the script will more likely contain the
up-to-date list of dependencies, because we use it in the CI, in
contrast to the list of dependencies in the testing.md which we might
forget to update).